# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "REST API server for managing Kibaship Operator resources",
    "title": "Kibaship Operator API",
    "termsOfService": "https://github.com/kibamail/kibaship",
    "contact": {
      "name": "API Support",
      "email": "support@kibamail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0"
  },
  "host": "localhost:8080",
  "basePath": "/v1",
  "paths": {
...
    "/v1/applications/{uuid}/deployments": {
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Retrieve all deployments for a specific application",
        "produces": ["application/json"],
        "tags": ["deployments"],
        "summary": "Get deployments by application",
        "parameters": [
          {
            "type": "string",
            "description": "Application UUID or slug (8-character identifier)",
            "name": "uuid",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "List of deployments",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/models.DeploymentResponse"
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "schema": {
              "$ref": "#/definitions/auth.ErrorResponse"
            }
          },
          "404": {
            "description": "Application not found",
            "schema": {
              "$ref": "#/definitions/auth.ErrorResponse"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/auth.ErrorResponse"
            }
          }
        }
      },
      "post": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Create a new deployment for an application",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["deployments"],
        "summary": "Create a new deployment",
        "parameters": [
          {
            "type": "string",
            "description": "Application UUID or slug",
            "name": "uuid",
            "in": "path",
            "required": true
          },
          {
            "description": "Deployment creation data",
            "name": "deployment",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/models.DeploymentCreateRequest"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Deployment created successfully",
            "schema": {
              "$ref": "#/definitions/models.DeploymentResponse"
            }
          },
          "400": {
            "description": "Validation errors in request data",
            "schema": {
              "$ref": "#/definitions/models.ValidationErrors"
            }
          },
          "401": {
            "description": "Authentication required",
            "schema": {
              "$ref": "#/definitions/auth.ErrorResponse"
            }
          },
          "404": {
            "description": "Application not found",
            "schema": {
              "$ref": "#/definitions/auth.ErrorResponse"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/auth.ErrorResponse"
            }
          }
        }
      }
    },
...
    "/v1/deployments/{uuid}": {
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Retrieve a deployment by its unique UUID or slug identifier",
        "produces": ["application/json"],
        "tags": ["deployments"],
        "summary": "Get deployment by UUID",
        "parameters": [
          {
            "type": "string",
            "description": "Deployment UUID or slug (8-character identifier)",
            "name": "uuid",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Deployment details",
            "schema": {
              "$ref": "#/definitions/models.DeploymentResponse"
            }
          },
          "401": {
            "description": "Authentication required",
            "schema": {
              "$ref": "#/definitions/auth.ErrorResponse"
            }
          },
          "404": {
            "description": "Deployment not found",
            "schema": {
              "$ref": "#/definitions/auth.ErrorResponse"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/auth.ErrorResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "auth.ErrorResponse": {
      "type": "object",
      "properties": {
        "error": {
          "type": "string",
          "example": "Unauthorized"
        },
        "message": {
          "type": "string",
          "example": "Invalid API key"
        }
      }
    },
...
    "models.DeploymentCreateRequest": {
      "type": "object",
      "required": ["applicationUuid"],
      "properties": {
        "applicationUuid": {
          "type": "string",
          "example": "550e8400-e29b-41d4-a716-446655440001"
        },
        "gitRepository": {
          "$ref": "#/definitions/models.GitRepositoryDeploymentConfig"
        }
      }
    },
    "models.DeploymentPhase": {
      "type": "string",
      "enum": ["Initializing", "Running", "Succeeded", "Failed", "Waiting"],
      "x-enum-varnames": [
        "DeploymentPhaseInitializing",
        "DeploymentPhaseRunning",
        "DeploymentPhaseSucceeded",
        "DeploymentPhaseFailed",
        "DeploymentPhaseWaiting"
      ]
    },
    "models.DeploymentResponse": {
      "type": "object",
      "properties": {
        "applicationSlug": {
          "type": "string",
          "example": "abc123de"
        },
        "applicationUuid": {
          "type": "string",
          "example": "550e8400-e29b-41d4-a716-446655440001"
        },
        "createdAt": {
          "type": "string",
          "example": "2023-01-01T12:00:00Z"
        },
        "gitRepository": {
          "$ref": "#/definitions/models.GitRepositoryDeploymentConfig"
        },
        "phase": {
          "allOf": [
            {
              "$ref": "#/definitions/models.DeploymentPhase"
            }
          ],
          "example": "Initializing"
        },
        "projectUuid": {
          "type": "string",
          "example": "550e8400-e29b-41d4-a716-446655440002"
        },
        "slug": {
          "type": "string",
          "example": "def456gh"
        },
        "updatedAt": {
          "type": "string",
          "example": "2023-01-01T12:00:00Z"
        },
        "uuid": {
          "type": "string",
          "example": "550e8400-e29b-41d4-a716-446655440000"
        }
      }
    },
...
    "models.GitRepositoryDeploymentConfig": {
      "type": "object",
      "required": ["commitSHA"],
      "properties": {
        "branch": {
          "type": "string",
          "example": "main"
        },
        "commitSHA": {
          "type": "string",
          "example": "abc123def456"
        }
      }
    },
...
    "models.ValidationError": {
      "type": "object",
      "properties": {
        "field": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      }
    },
    "models.ValidationErrors": {
      "type": "object",
      "properties": {
        "errors": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/models.ValidationError"
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "BearerAuth": {
      "description": "Type \"Bearer\" followed by a space and JWT token.",
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}